Create separate modules for each level on API
